refactor: Room 엔티티 패키지 위치 변경 #62
balance/content -> balance/room
